+++ manager/general/src/main/java/org/apache/doris/stack/entity/StudioSettingEntityPk.java
@@ -30,5 +30,5 @@ public class StudioSettingEntityPk implements Serializable {
 
     private String key;
 
-    private int clusterId;
+    private long clusterId;
 }
